+++ Client/index.html
@@ -30,6 +30,6 @@
     <div class="list-area" id="completed-list">
     </div>
   </div>
-  <script src="functions.js"></script>
+  <script type = "module" src="functions.js"></script>
 </body>
 </html>
